refactor: prepare newVaccination to activity switch
