+++ src/components/patientsDatabase/styles/NewVaccination.style.ts
